feat: 필터 보여주기 방식 변경
